+++ src/labs/broker.py
@@ -2,7 +2,7 @@
 
 """
 
-from .config import settings
+from .settings import settings
 
 # See PyPI for more information about taskiq_redis
 # https://pypi.org/project/taskiq-redis/
